+++ src/features/character/data/chat-types.ts
@@ -1,0 +1,4 @@
+import { type conversations } from './convo.json'
+
+export type CharacterUser = (typeof conversations)[number]
+export type Convo = CharacterUser['messages'][number]