+++ src/features/users/data/data.ts
@@ -1,7 +1,7 @@
 import { Shield, UserCheck, Users, CreditCard } from 'lucide-react'
-import { type UserStatus } from './schema'
+import { type VaccineStatus } from './schema'
 
-export const callTypes = new Map<UserStatus, string>([
+export const callTypes = new Map<VaccineStatus, string>([
   ['active', 'bg-teal-100/30 text-teal-900 dark:text-teal-200 border-teal-200'],
   ['inactive', 'bg-neutral-300/40 border-neutral-300'],
   ['invited', 'bg-sky-200/40 text-sky-900 dark:text-sky-100 border-sky-300'],