+++ src/features/users/data/schema.ts
@@ -1,12 +1,12 @@
 import { z } from 'zod'
 
-const userStatusSchema = z.union([
+const vaccineStatusSchema = z.union([
   z.literal('active'),
   z.literal('inactive'),
   z.literal('invited'),
   z.literal('suspended'),
 ])
-export type UserStatus = z.infer<typeof userStatusSchema>
+export type VaccineStatus = z.infer<typeof vaccineStatusSchema>
 
 const userRoleSchema = z.union([
   z.literal('superadmin'),
@@ -22,7 +22,7 @@ const userSchema = z.object({
   username: z.string(),
   email: z.string(),
   phoneNumber: z.string(),
-  status: userStatusSchema,
+  status: vaccineStatusSchema,
   role: userRoleSchema,
   createdAt: z.coerce.date(),
   updatedAt: z.coerce.date(),