+++ src/routes/_authenticated/character/index.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from '@tanstack/react-router'
+import { Character } from '@/features/character'
+
+export const Route = createFileRoute('/_authenticated/character/')({
+  component: Character,
+})